requirements.txt
Flask==3.1.0
gunicorn
numpy
scipy
scikit-learn
Pillow
python-dotenv
requests
websockets
sib-api-v3-sdk
africastalking
